# realworld/api/friend/v1/friend_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v5.29.3
// source: friend/v1/friend.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationFriendServiceAcceptFriendRequest = "/friend.v1.FriendService/AcceptFriendRequest"
const OperationFriendServiceGetRanking = "/friend.v1.FriendService/GetRanking"
const OperationFriendServiceListFriendRequests = "/friend.v1.FriendService/ListFriendRequests"
const OperationFriendServiceListFriends = "/friend.v1.FriendService/ListFriends"
const OperationFriendServiceRejectFriendRequest = "/friend.v1.FriendService/RejectFriendRequest"
const OperationFriendServiceSearchUsers = "/friend.v1.FriendService/SearchUsers"
const OperationFriendServiceSendFriendRequest = "/friend.v1.FriendService/SendFriendRequest"

type FriendServiceHTTPServer interface {
	// AcceptFriendRequest POST /api/v1/friends/requests/{id}/accept 同意申请，path: id 为申请 ID
	AcceptFriendRequest(context.Context, *AcceptFriendRequestRequest) (*AcceptFriendRequestReply, error)
	// GetRanking GET /api/v1/ranking 排行榜
	GetRanking(context.Context, *GetRankingRequest) (*GetRankingReply, error)
	// ListFriendRequests GET /api/v1/friends/requests 收到的好友申请列表
	ListFriendRequests(context.Context, *ListFriendRequestsRequest) (*ListFriendRequestsReply, error)
	// ListFriends GET /api/v1/friends 我的好友列表
	ListFriends(context.Context, *ListFriendsRequest) (*ListFriendsReply, error)
	// RejectFriendRequest POST /api/v1/friends/requests/{id}/reject 拒绝申请
	RejectFriendRequest(context.Context, *RejectFriendRequestRequest) (*RejectFriendRequestReply, error)
	// SearchUsers GET /api/v1/friends/search 搜索用户
	SearchUsers(context.Context, *SearchUsersRequest) (*SearchUsersReply, error)
	// SendFriendRequest POST /api/v1/friends/request 发送好友申请
	SendFriendRequest(context.Context, *SendFriendRequestRequest) (*SendFriendRequestReply, error)
}

func RegisterFriendServiceHTTPServer(s *http.Server, srv FriendServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/friends", _FriendService_ListFriends0_HTTP_Handler(srv))
	r.GET("/api/v1/friends/search", _FriendService_SearchUsers0_HTTP_Handler(srv))
	r.POST("/api/v1/friends/request", _FriendService_SendFriendRequest0_HTTP_Handler(srv))
	r.GET("/api/v1/friends/requests", _FriendService_ListFriendRequests0_HTTP_Handler(srv))
	r.POST("/api/v1/friends/requests/{id}/accept", _FriendService_AcceptFriendRequest0_HTTP_Handler(srv))
	r.POST("/api/v1/friends/requests/{id}/reject", _FriendService_RejectFriendRequest0_HTTP_Handler(srv))
	r.GET("/api/v1/ranking", _FriendService_GetRanking0_HTTP_Handler(srv))
}

func _FriendService_ListFriends0_HTTP_Handler(srv FriendServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListFriendsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFriendServiceListFriends)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListFriends(ctx, req.(*ListFriendsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListFriendsReply)
		return ctx.Result(200, reply)
	}
}

func _FriendService_SearchUsers0_HTTP_Handler(srv FriendServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SearchUsersRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFriendServiceSearchUsers)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SearchUsers(ctx, req.(*SearchUsersRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SearchUsersReply)
		return ctx.Result(200, reply)
	}
}

func _FriendService_SendFriendRequest0_HTTP_Handler(srv FriendServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SendFriendRequestRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFriendServiceSendFriendRequest)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SendFriendRequest(ctx, req.(*SendFriendRequestRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SendFriendRequestReply)
		return ctx.Result(200, reply)
	}
}

func _FriendService_ListFriendRequests0_HTTP_Handler(srv FriendServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListFriendRequestsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFriendServiceListFriendRequests)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListFriendRequests(ctx, req.(*ListFriendRequestsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListFriendRequestsReply)
		return ctx.Result(200, reply)
	}
}

func _FriendService_AcceptFriendRequest0_HTTP_Handler(srv FriendServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in AcceptFriendRequestRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFriendServiceAcceptFriendRequest)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.AcceptFriendRequest(ctx, req.(*AcceptFriendRequestRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*AcceptFriendRequestReply)
		return ctx.Result(200, reply)
	}
}

func _FriendService_RejectFriendRequest0_HTTP_Handler(srv FriendServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RejectFriendRequestRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFriendServiceRejectFriendRequest)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RejectFriendRequest(ctx, req.(*RejectFriendRequestRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RejectFriendRequestReply)
		return ctx.Result(200, reply)
	}
}

func _FriendService_GetRanking0_HTTP_Handler(srv FriendServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetRankingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationFriendServiceGetRanking)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetRanking(ctx, req.(*GetRankingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetRankingReply)
		return ctx.Result(200, reply)
	}
}

type FriendServiceHTTPClient interface {
	// AcceptFriendRequest POST /api/v1/friends/requests/{id}/accept 同意申请，path: id 为申请 ID
	AcceptFriendRequest(ctx context.Context, req *AcceptFriendRequestRequest, opts ...http.CallOption) (rsp *AcceptFriendRequestReply, err error)
	// GetRanking GET /api/v1/ranking 排行榜
	GetRanking(ctx context.Context, req *GetRankingRequest, opts ...http.CallOption) (rsp *GetRankingReply, err error)
	// ListFriendRequests GET /api/v1/friends/requests 收到的好友申请列表
	ListFriendRequests(ctx context.Context, req *ListFriendRequestsRequest, opts ...http.CallOption) (rsp *ListFriendRequestsReply, err error)
	// ListFriends GET /api/v1/friends 我的好友列表
	ListFriends(ctx context.Context, req *ListFriendsRequest, opts ...http.CallOption) (rsp *ListFriendsReply, err error)
	// RejectFriendRequest POST /api/v1/friends/requests/{id}/reject 拒绝申请
	RejectFriendRequest(ctx context.Context, req *RejectFriendRequestRequest, opts ...http.CallOption) (rsp *RejectFriendRequestReply, err error)
	// SearchUsers GET /api/v1/friends/search 搜索用户
	SearchUsers(ctx context.Context, req *SearchUsersRequest, opts ...http.CallOption) (rsp *SearchUsersReply, err error)
	// SendFriendRequest POST /api/v1/friends/request 发送好友申请
	SendFriendRequest(ctx context.Context, req *SendFriendRequestRequest, opts ...http.CallOption) (rsp *SendFriendRequestReply, err error)
}

type FriendServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewFriendServiceHTTPClient(client *http.Client) FriendServiceHTTPClient {
	return &FriendServiceHTTPClientImpl{client}
}

// AcceptFriendRequest POST /api/v1/friends/requests/{id}/accept 同意申请，path: id 为申请 ID
func (c *FriendServiceHTTPClientImpl) AcceptFriendRequest(ctx context.Context, in *AcceptFriendRequestRequest, opts ...http.CallOption) (*AcceptFriendRequestReply, error) {
	var out AcceptFriendRequestReply
	pattern := "/api/v1/friends/requests/{id}/accept"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationFriendServiceAcceptFriendRequest))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetRanking GET /api/v1/ranking 排行榜
func (c *FriendServiceHTTPClientImpl) GetRanking(ctx context.Context, in *GetRankingRequest, opts ...http.CallOption) (*GetRankingReply, error) {
	var out GetRankingReply
	pattern := "/api/v1/ranking"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationFriendServiceGetRanking))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListFriendRequests GET /api/v1/friends/requests 收到的好友申请列表
func (c *FriendServiceHTTPClientImpl) ListFriendRequests(ctx context.Context, in *ListFriendRequestsRequest, opts ...http.CallOption) (*ListFriendRequestsReply, error) {
	var out ListFriendRequestsReply
	pattern := "/api/v1/friends/requests"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationFriendServiceListFriendRequests))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListFriends GET /api/v1/friends 我的好友列表
func (c *FriendServiceHTTPClientImpl) ListFriends(ctx context.Context, in *ListFriendsRequest, opts ...http.CallOption) (*ListFriendsReply, error) {
	var out ListFriendsReply
	pattern := "/api/v1/friends"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationFriendServiceListFriends))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// RejectFriendRequest POST /api/v1/friends/requests/{id}/reject 拒绝申请
func (c *FriendServiceHTTPClientImpl) RejectFriendRequest(ctx context.Context, in *RejectFriendRequestRequest, opts ...http.CallOption) (*RejectFriendRequestReply, error) {
	var out RejectFriendRequestReply
	pattern := "/api/v1/friends/requests/{id}/reject"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationFriendServiceRejectFriendRequest))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// SearchUsers GET /api/v1/friends/search 搜索用户
func (c *FriendServiceHTTPClientImpl) SearchUsers(ctx context.Context, in *SearchUsersRequest, opts ...http.CallOption) (*SearchUsersReply, error) {
	var out SearchUsersReply
	pattern := "/api/v1/friends/search"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationFriendServiceSearchUsers))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// SendFriendRequest POST /api/v1/friends/request 发送好友申请
func (c *FriendServiceHTTPClientImpl) SendFriendRequest(ctx context.Context, in *SendFriendRequestRequest, opts ...http.CallOption) (*SendFriendRequestReply, error) {
	var out SendFriendRequestReply
	pattern := "/api/v1/friends/request"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationFriendServiceSendFriendRequest))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}