# realworld/api/inventory/v1/inventory_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v5.29.3
// source: inventory/v1/inventory.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationInventoryServiceList = "/inventory.v1.InventoryService/List"
const OperationInventoryServiceUseItem = "/inventory.v1.InventoryService/UseItem"

type InventoryServiceHTTPServer interface {
	// List GET /api/v1/inventory 我的背包
	List(context.Context, *ListRequest) (*ListReply, error)
	// UseItem POST /api/v1/inventory/use 使用道具
	UseItem(context.Context, *UseItemRequest) (*UseItemReply, error)
}

func RegisterInventoryServiceHTTPServer(s *http.Server, srv InventoryServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/inventory", _InventoryService_List0_HTTP_Handler(srv))
	r.POST("/api/v1/inventory/use", _InventoryService_UseItem0_HTTP_Handler(srv))
}

func _InventoryService_List0_HTTP_Handler(srv InventoryServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationInventoryServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*ListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListReply)
		return ctx.Result(200, reply)
	}
}

func _InventoryService_UseItem0_HTTP_Handler(srv InventoryServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UseItemRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationInventoryServiceUseItem)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UseItem(ctx, req.(*UseItemRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UseItemReply)
		return ctx.Result(200, reply)
	}
}

type InventoryServiceHTTPClient interface {
	// List GET /api/v1/inventory 我的背包
	List(ctx context.Context, req *ListRequest, opts ...http.CallOption) (rsp *ListReply, err error)
	// UseItem POST /api/v1/inventory/use 使用道具
	UseItem(ctx context.Context, req *UseItemRequest, opts ...http.CallOption) (rsp *UseItemReply, err error)
}

type InventoryServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewInventoryServiceHTTPClient(client *http.Client) InventoryServiceHTTPClient {
	return &InventoryServiceHTTPClientImpl{client}
}

// List GET /api/v1/inventory 我的背包
func (c *InventoryServiceHTTPClientImpl) List(ctx context.Context, in *ListRequest, opts ...http.CallOption) (*ListReply, error) {
	var out ListReply
	pattern := "/api/v1/inventory"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationInventoryServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UseItem POST /api/v1/inventory/use 使用道具
func (c *InventoryServiceHTTPClientImpl) UseItem(ctx context.Context, in *UseItemRequest, opts ...http.CallOption) (*UseItemReply, error) {
	var out UseItemReply
	pattern := "/api/v1/inventory/use"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationInventoryServiceUseItem))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}