# realworld/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/friends/request:
        post:
            tags:
                - FriendService
            description: POST /api/v1/friends/request 发送好友申请
            operationId: FriendService_SendFriendRequest
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/friend.v1.SendFriendRequestRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/friend.v1.SendFriendRequestReply'
...
    /api/v1/friends/requests/{id}/accept:
        post:
            tags:
                - FriendService
            description: 'POST /api/v1/friends/requests/{id}/accept 同意申请，path: id 为申请 ID'
            operationId: FriendService_AcceptFriendRequest
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/friend.v1.AcceptFriendRequestRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/friend.v1.AcceptFriendRequestReply'
    /api/v1/friends/requests/{id}/reject:
        post:
            tags:
                - FriendService
            description: POST /api/v1/friends/requests/{id}/reject 拒绝申请
            operationId: FriendService_RejectFriendRequest
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/friend.v1.RejectFriendRequestRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/friend.v1.RejectFriendRequestReply'
    /api/v1/friends/search:
        get:
            tags:
                - FriendService
            description: GET /api/v1/friends/search 搜索用户
            operationId: FriendService_SearchUsers
            parameters:
                - name: keyword
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/friend.v1.SearchUsersReply'
...
    /api/v1/inventory/use:
        post:
            tags:
                - InventoryService
            description: POST /api/v1/inventory/use 使用道具
            operationId: InventoryService_UseItem
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/inventory.v1.UseItemRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/inventory.v1.UseItemReply'
...
    /api/v1/ranking:
        get:
            tags:
                - FriendService
            description: GET /api/v1/ranking 排行榜
            operationId: FriendService_GetRanking
            parameters:
                - name: type
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/friend.v1.GetRankingReply'
...
components:
    schemas:
...
        friend.v1.AcceptFriendRequestReply:
            type: object
            properties:
                success:
                    type: boolean
                message:
                    type: string
        friend.v1.AcceptFriendRequestRequest:
            type: object
            properties:
                id:
                    type: string
            description: 'AcceptFriendRequest 同意申请，path: id 为申请 ID'
...
        friend.v1.GetRankingReply:
            type: object
            properties:
                ranking:
                    type: array
                    items:
                        $ref: '#/components/schemas/friend.v1.UserBrief'
...
        friend.v1.RejectFriendRequestReply:
            type: object
            properties:
                success:
                    type: boolean
                message:
                    type: string
        friend.v1.RejectFriendRequestRequest:
            type: object
            properties:
                id:
                    type: string
            description: RejectFriendRequest 拒绝申请
        friend.v1.SearchUsersReply:
            type: object
            properties:
                users:
                    type: array
                    items:
                        $ref: '#/components/schemas/friend.v1.UserBrief'
        friend.v1.SendFriendRequestReply:
            type: object
            properties:
                success:
                    type: boolean
                message:
                    type: string
        friend.v1.SendFriendRequestRequest:
            type: object
            properties:
                to_user_id:
                    type: string
                message:
                    type: string
            description: SendFriendRequest 发送好友申请
        friend.v1.UserBrief:
            type: object
            properties:
                id:
                    type: string
                username:
                    type: string
                nickname:
                    type: string
                avatar_url:
                    type: string
                level:
                    type: integer
                    format: int32
                heart_points:
                    type: integer
                    format: int32
            description: UserBrief 用户简要信息（搜索、好友列表、排行榜共用）
...
        inventory.v1.UseItemReply:
            type: object
            properties:
                success:
                    type: boolean
                message:
                    type: string
        inventory.v1.UseItemRequest:
            type: object
            properties:
                item_id:
                    type: string
                pet_id:
                    type: string
            description: UseItem 使用道具，如喂食需传 pet_id
...